Add node default ssl certs and fallback
In the future our certificates will expire and we want to allow clients
to continue to work without needing to upgrade. We also want to expire
old certificates from the options bundled with the client.

This update uses the default node certificates by default and will
fallback to bundled certs in case the node defaults don't work or don't
exist.

Update log method to print more reasonable log messages since it was
abstracted. Using `arguments` lead to an object with strings of numbers
as keys.

This tracks success of the tls connection by verifying the data send
happens within a reasonable amount of time. If that doesn't happen it
will try a different set of tls options.

This also updates the tests to do additional verification based on log
messages.