Fix tests since the recent certificate expiration has passed
Previously the tests relied on the old ELB certificate setup, so after
we switched the certs some tests failed. This comments out some tests
that don't make sense now that we don't have an ELB with old certs. It
also updates some tests the were relying on the current time to now
explicitly specify the time before or after cert expiration in the
tests. This also adds a "fallback" test to cover the "after expiration"
case.