DRY up test config
Previously we had copied some test config around, but during recent
changes we noticed an issue with the debug option that was fixed in one
place but not others. We also needed to copy more test setup code.

This seemed like a reasonable time to bite the bullet since we found an
issue and were about to make the problem worse.